Replace gammaCorrection hack with sRGB color interpolation
TIL that SVG filters default to “linearRGB”, which produces washed-
out colors (for example, see `test/water-output-incorrect.jpg`).

We’ve been using a `gammaCorrection` hack to convert linearRGB
values to sRGB, which raised each value to the power of `1/2.2`
(or `0.45` for short).

While this worked, it’s neither the simplest nor the most accurate
solution. Now that I understand the problem, fixing it simply a
matter of adding `color-interpolation-filters="sRGB"` to the
filter.

https://twitter.com/atannerhodges/status/1021101810919116802